+++ src/AssetMate.php
@@ -12,6 +12,7 @@ use craft\log\MonologTarget;
 
 use Psr\Log\LogLevel;
 
+use vaersaagod\assetmate\helpers\CpHelper;
 use vaersaagod\assetmate\models\Settings;
 use vaersaagod\assetmate\services\Convert;
 use vaersaagod\assetmate\services\Resize;
@@ -94,13 +95,15 @@ class AssetMate extends Plugin
             }
         );
 
-//        Event::on(
-//            Asset::class,
-//            Element::EVENT_DEFINE_META_FIELDS_HTML,
-//            static function (DefineHtmlEvent $event) {
-//                $event->html .= '<div>Hei</div>';
-//            }
-//        );
+        Event::on(
+            Asset::class,
+            Element::EVENT_DEFINE_SIDEBAR_HTML,
+            static function (DefineHtmlEvent $event) {
+                /** @var Asset $asset */
+                $asset = $event->sender;
+                $event->html .= CpHelper::renderAssetMetaDataHtml($asset);
+            }
+        );
 
     }